pyproject.toml
[project]
name = "lakepulse"
version = "0.1.0"
description = "Real-time Wikipedia edits streamed through Databricks"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31",
    "databricks-sdk>=0.30",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "psycopg[binary]>=3.2",
    "pydantic>=2.0",
    "alembic>=1.18.4",
    "sqlalchemy>=2.0.49",
    "pydantic-settings>=2.13.1",
]

[project.optional-dependencies]
collector = [
    "databricks-zerobus-ingest-sdk>=0.1",
]

[project.scripts]
lakepulse-collect = "collector.main:main"
